requirements.txt
streamlit
requests
python-dotenv
plotly
geocoder

